requirements.txt
tqdm
psutil
pathlib
numpy==1.17.4
opencv-python==4.1.2.30
scikit-image
Pillow==6.2.1
scikit-learn
toposort
fastcluster
matplotlib==3.1.1
imageio==2.6.1
imageio-ffmpeg
ffmpy==0.2.2
# Revert back to nvidia-ml-py3 when windows/system32 patch is implemented
git+https://github.com/deepfakes/nvidia-ml-py3.git
#nvidia-ml-py3
h5py==2.9.0
Keras==2.2.4
pywin32 ; sys_platform == "win32"
pynvx==1.0.0 ; sys_platform == "darwin"

# tensorflow is included within the docker image.
# If you are looking for dependencies for a manual install,

# NB: Tensorflow version 1.12 is the minimum supported version of Tensorflow.
#    If your graphics card support is below Cuda 9.0 you will need to either
#    compile tensorflow yourself or download a custom version.
#    Install 1.12.0<=tensorflow-gpu<=1.13.0 for CUDA 9.0
#    or 1.13.1<=tensorflow-gpu<1.15 for CUDA 10.0
